se sube

+++ Proyecto/Domain/Model/Factura.cs
@@ -7,12 +7,12 @@
         private string tipodePago =null;
         private string documentoCliente=null;
         private string nombreCliente = null;
-        private int subtotal = 0;
-        private int descuento = 0;
-        private int iva = 0;
-        private int totalDescuento = 0;
-        private int totalImpuesto = 0;
-        private int total = 0;
+        private decimal subtotal = 0;
+        private decimal descuento = 0;
+        private decimal iva = 0;
+        private decimal totalDescuento = 0;
+        private decimal totalImpuesto = 0;
+        private decimal total = 0;
        
     
         public int Id { get; set; }
@@ -23,13 +23,13 @@
         public string TipodePago { get { return tipodePago; } set { tipodePago = value; } }
         public string DocumentoCliente { get { return documentoCliente; } set { documentoCliente = value; } }
         public string NombreCliente { get { return nombreCliente; } set { nombreCliente = value; } }
-        public int Subtotal { get { return subtotal; } set { subtotal = value; } }
-        public int Descuento { get { return descuento; } set { descuento = value; } }
-        public int Iva { get { return iva; } set { iva = value; } }
+        public decimal Subtotal { get { return subtotal; } set { subtotal = value; } }
+        public decimal Descuento { get { return descuento; } set { descuento = value; } }
+        public decimal Iva { get { return iva; } set { iva = value; } }
 
-        public int TotalDescuento { get { return totalDescuento; } set { totalDescuento = value; } }
-        public int TotalImpuesto { get { return totalImpuesto; } set { totalImpuesto = value; } }
-        public int Total { get { return total; } set { total = value; } }
+        public decimal TotalDescuento { get { return totalDescuento; } set { totalDescuento = value; } }
+        public decimal TotalImpuesto { get { return totalImpuesto; } set { totalImpuesto = value; } }
+        public decimal Total { get { return total; } set { total = value; } }
